Refactor: separate manage permissions action saga
